Work Session 1: 12/16/20 9:10pm - 10:20pm
In this work session, I was able to complete two of the problems. These problems were easy for me to complete with some minor issues with how I presented them at the end. However, I am happy that I completed these problems in the 70min time slot and found the problems interesting and fun. Overall, I like how the questions are simple and am proud that I made sense of them.