Work Session 2: 1/6/20 11:50pm - 1/7/2 1:20am
I studied other people's answers to this problem because I spend too much time-solving it. I also read the question wrong so I couldn't reach an answer. I didn't know how to find the prime factors without making a variable in a range of numbers and then solving what applies for each number. When I searched for the answer, I also had a hard time comprehending, so I made comments to explain. I learned how the largest prime number cannot be greater than the square root of the number because it is the greatest two factors that give the product value.