Work Session 4: 1/13/20 11:20pm - 12:10 am
In this work session, I finished the previous problem for the largest palindrome. However, there is a sorting issue - I don't know if it is the code or the system itself because when I put the command "sort(list)" it doesn't sort it correctly. If it runs correctly, the output would be correct. The way I approached this was I first thought of how to identify a palindrome and try all the combinations of third digits numbers. I made a list to add all the palindromes the two three-digit numbers created and see if the correct answer was on it. It was, however, the list wasn't sorted correctly so I couldn't get that specific number as an output. After I finished this problem, I moved on to the next one and wrote down an approach. Hopefully, I can finish it next session.